+++ src/main/java/com/dsm/dao/ICartDao.java
@@ -34,7 +34,7 @@ public interface ICartDao {
      *
      * @param cartItemId 购物车一项的唯一标识id
      */
-    long deleteCartItem(int cartItemId);
+    long deleteCartItem(long cartItemId);
 
 
     /**
@@ -42,7 +42,7 @@ public interface ICartDao {
      *
      * @param userId 用户ID
      */
-    long cleanCartAll(int userId);
+    long cleanCartAll(long userId);
 
     /**
      * 根据cartItemId 获取购物车中的商品信息
@@ -50,13 +50,13 @@ public interface ICartDao {
      * @param cartItemId 购物车一项的唯一标识id
      * @return 购物车一条子项
      */
-    ShoppingCartItemPO getShoppingCartItem(int cartItemId);
+    ShoppingCartItemPO getShoppingCartItem(long cartItemId);
 
 
     /**
      * 根据商品的skuId和用户ID 获取购物车项
      */
-    ShoppingCartItemPO getShoppingCartItemBySkuId(@Param("userId")int userId, @Param("skuId")int skuId);
+    ShoppingCartItemPO getShoppingCartItemBySkuId(@Param("userId")long userId, @Param("skuId")long skuId);
 
 
     /**
@@ -66,7 +66,7 @@ public interface ICartDao {
      * @param shopId 店铺ID
      * @return 购物车商品信息列表
      */
-    List<ShoppingCartItemPO> getShoppingCartItemByShopId(@Param("userId") int userId, @Param("shopId") int shopId);
+    List<ShoppingCartItemPO> getShoppingCartItemByShopId(@Param("userId") long userId, @Param("shopId") long shopId);
 
 
     /**
@@ -75,7 +75,7 @@ public interface ICartDao {
      * @param userId 用户ID
      * @return 购物车所有子项
      */
-    List<ShoppingCartItemPO> getShoppingCartInfoAll(int userId);
+    List<ShoppingCartItemPO> getShoppingCartInfoAll(long userId);
 
 
 }
